requirements.txt
fastapi
uvicorn
sqlalchemy
passlib
bcrypt
pydantic
python-jose
python-multipart
psycopg2-binary